changed the layout of the files so that fixtures are global* to other pytests
