+++ backend/money_maker/auth/routes.py
@@ -5,7 +5,6 @@ from flask import Blueprint, jsonify, request
 from flask_jwt_extended import (create_access_token, get_jwt, get_jwt_identity,
                                 jwt_required, set_access_cookies,
                                 unset_jwt_cookies)
-
 from money_maker.extensions import bcrypt, db, jwt_manager
 from money_maker.models.user import User, users_schema
 
@@ -59,6 +58,14 @@ def login():
 
 @auth_bp.route("/logout", methods=["POST"])
 def logout():
+    """
+    Logs out a user from the frontend. Note that
+    a jwt is not required as potentially jwt's may be expired
+    or non-existant.
+
+    :return: The flask reponse
+    :rtype: flask.Response
+    """
     response = jsonify({"msg": "logout successful"})
     unset_jwt_cookies(response)
     return response
